requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
stripe==9.9.0
starlette==0.37.2
python-dotenv==1.0.1
pydantic==2.7.1
itsdangerous==2.2.0    
twilio